storeApp/requirements.txt
aiosmtplib==3.0.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
bcrypt==3.2.2
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2==2.9.9
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
